+++ src/pages/study/Create.jsx
@@ -7,28 +7,6 @@ import StudyPreview from '@components/pages/study/create/StudyPreview';
 import { useState } from 'react';
 
 const Create = () => {
-  // 카테고리 더미 데이터(상수 처리 vs DB 호출 고려 중)
-  const BookCategoryList = [
-    { id: 1, title: '자기계발' },
-    { id: 2, title: '인문' },
-    { id: 3, title: '경제/경영' },
-    { id: 4, title: '처세' },
-    { id: 5, title: 'IT' },
-    { id: 6, title: '소설' },
-    { id: 7, title: '과학' },
-    { id: 8, title: '시/에세이' },
-    { id: 9, title: '역사/문화' },
-    { id: 10, title: '건강' },
-    { id: 11, title: '요리' },
-  ];
-
-  // 카테고리 드롭다운(추가 수정 예정)
-  const BookCategoryOption = BookCategoryList.map((item) => (
-    <option id={item.title} key={item.id} value={item.id}>
-      {item.title}
-    </option>
-  ));
-
   // 현재 작성 중인 step 상태로 지정
   const [currentStep, setCurrentStep] = useState(0);
 
@@ -73,9 +51,7 @@ const Create = () => {
               setCurrentStep={setCurrentStep}
             />
             {currentStep === 0 && <SearchBook />}
-            {currentStep === 1 && (
-              <StudyForm BookCategoryOption={BookCategoryOption} />
-            )}
+            {currentStep === 1 && <StudyForm />}
             {currentStep === 2 && <StudyPreview />}
             <div className="flex justify-between">
               {currentStep === 0 ? (
